pyproject.toml
[project]
name = "gha_runner"
version = "0.4.0"
description = "A GitHub Runner for deploying on cloud backends"
authors = [{ name = "Ethan Holz", email = "ethan.holz@omsf.io" }, {name = "David W.H. Swenson", email = "david.swenson@omsf.io"}]
readme = "README.md"
requires-python = ">= 3.11"
dependencies = ["boto3", "requests"]

[project.optional-dependencies]
test = [
    "pytest",
    "pytest-cov",
    "moto[ec2]",
    "responses"
]
convert = [
    "ruamel.yaml",
    "click",
]

[build-system]
# We are not going to add versioningit at this time
requires = ["setuptools >= 61.2"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
include = ["gha_runner", "gha_runner_convert"]

[tool.setuptools.package-data]
gha_runner = ["*.templ", "templates/*.templ"]

[tool.coverage.report]
exclude_also = ["raise NotImplementedError"]

[tool.pytest.ini_options]
markers = ["slow: marks test as slow","main: marks the test for the main module"]

[tool.ruff]
line-length = 80

[tool.ruff.lint]
select = ["F", "E", "N"]

